+++ R/chord_data.R
@@ -1,0 +1,26 @@
+#' @title Chord data as example data for the chord_plot function.
+#' @description Chord data as example data for the chord_plot function.
+#' @author benben-miao
+#'
+#' @docType data
+#' @format Dataframe: include columns (Samples, Set1, Set2, ...), rows (Gene1, Gene2, ...).
+#' @usage data(chord_data)
+#'
+#' @keywords datasets
+#'
+#' @references https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/ChordPlot/
+#'
+#' @examples
+#' # 1. Library TOmicsVis package
+#' library(TOmicsVis)
+#'
+#' # 2. Load example data
+#' data(chord_data)
+#'
+#' # 3. View example data
+#' chord_data
+#'
+"chord_data"
+
+# chord_data <- data
+# usethis::use_data(chord_data)