+++ R/circos_heatmap_data.R
@@ -1,0 +1,26 @@
+#' @title Gene expression as example data for circos_heatmap function.
+#' @description Gene expression as example data for circos_heatmap function.
+#' @author benben-miao
+#'
+#' @docType data
+#' @format Dataframe: include columns (Sample1, Sample2, ...), rows (Gene1, Gene2, ...).
+#' @usage data(circos_heatmap_data)
+#'
+#' @keywords datasets
+#'
+#' @references https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/CircosHeatmap/
+#'
+#' @examples
+#' # 1. Library TOmicsVis package
+#' library(TOmicsVis)
+#'
+#' # 2. Load example data
+#' data(circos_heatmap_data)
+#'
+#' # 3. View example data
+#' circos_heatmap_data
+#'
+"circos_heatmap_data"
+
+# circos_heatmap_data <- data
+# usethis::use_data(circos_heatmap_data)